CR fixes for https://github.com/metasfresh/metasfresh/pull/18322
