+++ backend/de.metas.contracts/src/main/java/de/metas/contracts/interceptor/C_Invoice.java
@@ -1,33 +1,16 @@
 package de.metas.contracts.interceptor;
 
-import de.metas.contracts.FlatrateTermId;
-import de.metas.contracts.IFlatrateBL;
 import de.metas.contracts.impl.CustomerRetentionRepository;
-import de.metas.contracts.modular.ComputingMethodType;
 import de.metas.contracts.modular.log.ModularContractLogService;
 import de.metas.document.DocTypeId;
-import de.metas.document.IDocTypeBL;
 import de.metas.invoice.InvoiceId;
-import de.metas.invoice.service.IInvoiceBL;
-import de.metas.invoicecandidate.api.IInvoiceCandDAO;
-import de.metas.invoicecandidate.model.I_C_Invoice_Candidate;
-import de.metas.util.Services;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
 import org.adempiere.ad.modelvalidator.annotations.DocValidate;
 import org.adempiere.ad.modelvalidator.annotations.Interceptor;
-import org.adempiere.exceptions.AdempiereException;
-import org.adempiere.model.InterfaceWrapperHelper;
-import org.compiere.SpringContextHolder;
 import org.compiere.model.I_C_Invoice;
-import org.compiere.model.I_C_InvoiceLine;
 import org.compiere.model.ModelValidator;
 import org.springframework.stereotype.Component;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 /*
  * #%L
@@ -56,70 +39,20 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class C_Invoice
 {
-	private final IDocTypeBL docTypeBL = Services.get(IDocTypeBL.class);
-	private final IInvoiceBL invoiceBL = Services.get(IInvoiceBL.class);
-	private final IFlatrateBL flatrateBL = Services.get(IFlatrateBL.class);
-	private final IInvoiceCandDAO invoiceCandDAO = Services.get(IInvoiceCandDAO.class);
 	@NonNull private final ModularContractLogService modularContractLogService;
+	@NonNull private final CustomerRetentionRepository customerRetentionRepo;
 
 	@DocValidate(timings = { ModelValidator.TIMING_AFTER_COMPLETE })
 	public void updateCustomerRetention(final I_C_Invoice invoice)
 	{
-		final CustomerRetentionRepository customerRetentionRepo = SpringContextHolder.instance.getBean(CustomerRetentionRepository.class);
 		final InvoiceId invoiceId = InvoiceId.ofRepoId(invoice.getC_Invoice_ID());
 
 		customerRetentionRepo.updateCustomerRetentionOnInvoiceComplete(invoiceId);
-
 	}
 
 	@DocValidate(timings = { ModelValidator.TIMING_AFTER_REVERSECORRECT, ModelValidator.TIMING_AFTER_VOID })
 	public void unprocessModularContractLogs(final I_C_Invoice invoice)
 	{
-		final DocTypeId docTypeId = DocTypeId.ofRepoId(invoice.getC_DocType_ID());
-		if ((!docTypeBL.isFinalInvoiceOrFinalCreditMemo(docTypeId)
-				&& !docTypeBL.isDefinitiveInvoiceOrDefinitiveCreditMemo(docTypeId))
-				&& !docTypeBL.isInterimInvoice(docTypeId))
-		{
-			return;
-		}
-
-		final InvoiceId invoiceId = InvoiceId.ofRepoId(invoice.getC_Invoice_ID());
-
-		modularContractLogService.unprocessLogsForInvoice(invoiceId, getComputingMethodTypes(docTypeId));
-
-		final List<I_C_Invoice_Candidate> invoiceCandidates = invoiceCandDAO.retrieveInvoiceCandidates(invoiceId);
-		for (final I_C_Invoice_Candidate invoiceCandidate : invoiceCandidates)
-		{
-			invoiceCandidate.setProcessed(false);//maybe should live in de.metas.invoicecandidate.api.impl.InvoiceCandBL.handleReversalForInvoice, but I'm afraid of consequences
-			invoiceCandidate.setIsActive(false);
-		}
-		InterfaceWrapperHelper.saveAll(invoiceCandidates);
-		if (docTypeBL.isDefinitiveInvoiceOrDefinitiveCreditMemo(docTypeId))
-		{
-			final Set<FlatrateTermId> contractIds = invoiceBL.getLines(invoiceId)
-					.stream()
-					.map(I_C_InvoiceLine::getC_Flatrate_Term_ID)
-					.map(FlatrateTermId::ofRepoId)
-					.collect(Collectors.toSet());
-			flatrateBL.reverseDefinitiveInvoice(contractIds);
-		}
-	}
-
-	@NonNull
-	private Collection<ComputingMethodType> getComputingMethodTypes(@NonNull final DocTypeId docTypeId)
-	{
-		if (docTypeBL.isFinalInvoiceOrFinalCreditMemo(docTypeId))
-		{
-			return ComputingMethodType.FINAL_INVOICE_SPECIFIC_METHODS;
-		}
-		if (docTypeBL.isDefinitiveInvoiceOrDefinitiveCreditMemo(docTypeId))
-		{
-			return ComputingMethodType.DEFINITIVE_INVOICE_SPECIFIC_METHODS;
-		}
-		if (docTypeBL.isInterimInvoice(docTypeId))
-		{
-			return ComputingMethodType.INTERIM_INVOICE_SPECIFIC_METHODS;
-		}
-		throw new AdempiereException("Unexpected document type: " + docTypeId);
+		modularContractLogService.unprocessModularContractLogs(InvoiceId.ofRepoId(invoice.getC_Invoice_ID()), DocTypeId.ofRepoId(invoice.getC_DocType_ID()));
 	}
 }
